+++ Source/Serialization/BaseSerializers/UnionSerializer.swift
@@ -1,0 +1,9 @@
+//
+//  UnionSerializer.swift
+//
+//  Copyright © 2019 Nervos Foundation. All rights reserved.
+//
+
+import Foundation
+
+/// Not implemented and used yet